requirements.txt
alembic==1.13.3
amqp==5.2.0
APScheduler==3.10.4
async-timeout==4.0.3
Authlib==1.3.2
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crontab==1.0.1
cryptography==43.0.3
Flask==2.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RQ2==18.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
freezegun==1.5.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
MarkupSafe==3.0.2
oauthlib==3.2.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rq==1.10.0
rq-scheduler==0.13.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==1.4.46
stripe==11.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
URLObject==2.4.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.4
WTForms==3.2.1
zipp==3.20.2
